refactor app entry and thunk dispatch
